changed downsample type from double to int since we will only allow integer downsamples.

+++ Code/Algorithms/Segmentation/ctcSegmentColonFilter.h
@@ -82,7 +82,7 @@ protected:
 private:
 
   signed short m_constShortROSMarker;
-  double downsampleFactor;
+  long int downsampleFactor;
 
   SegmentColonFilter(Self&);   // intentionally not implemented
   void operator=(const Self&);          // intentionally not implemented
